+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Update `<DaoDataListItem.Structure />` module component to support new `isExternal` property
+
 ### Changed
 
 - Update Manrope font weight for Regular variant from 500 to 400